Add Variable/Evaluate/Scope Request Handler
