+++ org.eclipse.jdt.ls.debug/src/org/eclipse/jdt/ls/debug/adapter/DebugAdapter.java
@@ -12,70 +12,32 @@
 package org.eclipse.jdt.ls.debug.adapter;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.function.BiConsumer;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
-import org.apache.commons.lang3.StringUtils;
-import org.eclipse.jdt.ls.debug.adapter.Messages.Response;
 import org.eclipse.jdt.ls.debug.adapter.Requests.Arguments;
 import org.eclipse.jdt.ls.debug.adapter.Requests.Command;
-import org.eclipse.jdt.ls.debug.adapter.Requests.EvaluateArguments;
-import org.eclipse.jdt.ls.debug.adapter.Requests.ScopesArguments;
-import org.eclipse.jdt.ls.debug.adapter.Requests.SetVariableArguments;
-import org.eclipse.jdt.ls.debug.adapter.Requests.VariablesArguments;
-import org.eclipse.jdt.ls.debug.adapter.formatter.NumericFormatEnum;
-import org.eclipse.jdt.ls.debug.adapter.formatter.NumericFormatter;
-import org.eclipse.jdt.ls.debug.adapter.formatter.SimpleTypeFormatter;
 import org.eclipse.jdt.ls.debug.adapter.handler.AttachRequestHandler;
 import org.eclipse.jdt.ls.debug.adapter.handler.ConfigurationDoneRequestHandler;
 import org.eclipse.jdt.ls.debug.adapter.handler.DisconnectRequestHandler;
+import org.eclipse.jdt.ls.debug.adapter.handler.EvaluateRequestHandler;
 import org.eclipse.jdt.ls.debug.adapter.handler.InitializeRequestHandler;
 import org.eclipse.jdt.ls.debug.adapter.handler.LaunchRequestHandler;
+import org.eclipse.jdt.ls.debug.adapter.handler.ScopesRequestHandler;
 import org.eclipse.jdt.ls.debug.adapter.handler.SetBreakpointsRequestHandler;
 import org.eclipse.jdt.ls.debug.adapter.handler.SetExceptionBreakpointsRequestHandler;
+import org.eclipse.jdt.ls.debug.adapter.handler.SetVariableRequestHandler;
 import org.eclipse.jdt.ls.debug.adapter.handler.SourceRequestHandler;
 import org.eclipse.jdt.ls.debug.adapter.handler.StackTraceRequestHandler;
 import org.eclipse.jdt.ls.debug.adapter.handler.ThreadsRequestHandler;
-import org.eclipse.jdt.ls.debug.adapter.variables.IVariableFormatter;
-import org.eclipse.jdt.ls.debug.adapter.variables.JdiObjectProxy;
-import org.eclipse.jdt.ls.debug.adapter.variables.StackFrameScope;
-import org.eclipse.jdt.ls.debug.adapter.variables.ThreadObjectReference;
-import org.eclipse.jdt.ls.debug.adapter.variables.Variable;
-import org.eclipse.jdt.ls.debug.adapter.variables.VariableFormatterFactory;
-import org.eclipse.jdt.ls.debug.adapter.variables.VariableUtils;
+import org.eclipse.jdt.ls.debug.adapter.handler.VariablesRequestHandler;
 import org.eclipse.jdt.ls.debug.internal.Logger;
-
-import com.sun.jdi.AbsentInformationException;
-import com.sun.jdi.ArrayReference;
-import com.sun.jdi.ArrayType;
-import com.sun.jdi.ClassNotLoadedException;
-import com.sun.jdi.ClassType;
-import com.sun.jdi.Field;
-import com.sun.jdi.InvalidTypeException;
-import com.sun.jdi.LocalVariable;
-import com.sun.jdi.ObjectReference;
-import com.sun.jdi.PrimitiveValue;
-import com.sun.jdi.ReferenceType;
-import com.sun.jdi.StackFrame;
-import com.sun.jdi.ThreadReference;
-import com.sun.jdi.Type;
-import com.sun.jdi.TypeComponent;
-import com.sun.jdi.Value;
 
 public class DebugAdapter implements IDebugAdapter {
     private BiConsumer<Events.DebugEvent, Boolean> eventConsumer;
-
     private IProviderContext providerContext;
-    private VariableRequestHandler variableRequestHandler;
-
     private IDebugAdapterContext debugContext = null;
     private Map<Command, List<IDebugRequestHandler>> requestHandlers = null;
 
@@ -85,7 +47,6 @@ public class DebugAdapter implements IDebugAdapter {
     public DebugAdapter(BiConsumer<Events.DebugEvent, Boolean> consumer, IProviderContext providerContext) {
         this.eventConsumer = consumer;
         this.providerContext = providerContext;
-        this.variableRequestHandler = new VariableRequestHandler(VariableFormatterFactory.createVariableFormatter());
         this.debugContext = new DebugAdapterContext(this);
         this.requestHandlers = new HashMap<>();
         initialize();
@@ -102,57 +63,14 @@ public class DebugAdapter implements IDebugAdapter {
         Arguments cmdArgs = JsonUtils.fromJson(request.arguments, command.getArgumentType());
 
         try {
-            switch (command) {
-                case SCOPES:
-                    scopes((ScopesArguments) cmdArgs, response);
-                    break;
-
-                case VARIABLES:
-                    Requests.VariablesArguments varArguments = (VariablesArguments) cmdArgs;
-                    if (varArguments.variablesReference == -1) {
-                        AdapterUtils.setErrorResponse(response, ErrorCode.ARGUMENT_MISSING,
-                                "VariablesRequest: property 'variablesReference' is missing, null, or empty");
-                    } else {
-                        variables(varArguments, response);
-                    }
-                    break;
-
-                case SETVARIABLE:
-                    Requests.SetVariableArguments setVarArguments = (SetVariableArguments) cmdArgs;
-                    if (setVarArguments.value == null) {
-                        // Just exit out of editing if we're given an empty expression.
-                        response.body = new Responses.ResponseBody();
-                    } else if (setVarArguments.variablesReference == -1) {
-                        AdapterUtils.setErrorResponse(response, ErrorCode.ARGUMENT_MISSING,
-                                "SetVariablesRequest: property 'variablesReference' is missing, null, or empty");
-                    } else if (setVarArguments.name == null) {
-                        AdapterUtils.setErrorResponse(response, ErrorCode.ARGUMENT_MISSING,
-                                "SetVariablesRequest: property 'name' is missing, null, or empty");
-                    } else {
-                        setVariable(setVarArguments, response);
-                    }
-                    break;
-
-                case EVALUATE:
-                    Requests.EvaluateArguments evaluateArguments = (EvaluateArguments) cmdArgs;
-                    if (evaluateArguments.expression == null) {
-                        AdapterUtils.setErrorResponse(response, ErrorCode.ARGUMENT_MISSING,
-                                "EvaluateRequest: property 'expression' is missing, null, or empty");
-                    } else {
-                        evaluate(evaluateArguments, response);
-                    }
-                    break;
-
-                default:
-                    List<IDebugRequestHandler> handlers = requestHandlers.get(command);
-                    if (handlers != null && !handlers.isEmpty()) {
-                        for (IDebugRequestHandler handler : handlers) {
-                            handler.handle(command, cmdArgs, response, this.debugContext);
-                        }
-                    } else {
-                        AdapterUtils.setErrorResponse(response, ErrorCode.UNRECOGNIZED_REQUEST_FAILURE,
-                                String.format("Unrecognized request: { _request: %s }", request.command));
-                    }
+            List<IDebugRequestHandler> handlers = requestHandlers.get(command);
+            if (handlers != null && !handlers.isEmpty()) {
+                for (IDebugRequestHandler handler : handlers) {
+                    handler.handle(command, cmdArgs, response, this.debugContext);
+                }
+            } else {
+                AdapterUtils.setErrorResponse(response, ErrorCode.UNRECOGNIZED_REQUEST_FAILURE,
+                        String.format("Unrecognized request: { _request: %s }", request.command));
             }
         } catch (Exception e) {
             Logger.logException("DebugSession dispatch exception", e);
@@ -197,6 +115,10 @@ public class DebugAdapter implements IDebugAdapter {
         registerHandler(new SourceRequestHandler());
         registerHandler(new ThreadsRequestHandler());
         registerHandler(new StackTraceRequestHandler());
+        registerHandler(new ScopesRequestHandler());
+        registerHandler(new VariablesRequestHandler());
+        registerHandler(new SetVariableRequestHandler());
+        registerHandler(new EvaluateRequestHandler());
     }
 
     private void registerHandler(IDebugRequestHandler handler) {
@@ -208,496 +130,5 @@ public class DebugAdapter implements IDebugAdapter {
             }
             handlerList.add(handler);
         }
-    }
-
-    /* ======================================================*/
-    /* Invoke different dispatch logic for different request */
-    /* ======================================================*/
-
-    private void scopes(Requests.ScopesArguments arguments, Response response) {
-        response.body = this.variableRequestHandler.scopes(arguments);
-    }
-
-    private void variables(Requests.VariablesArguments arguments, Response response) {
-        try {
-            response.body = this.variableRequestHandler.variables(arguments);
-        } catch (AbsentInformationException e) {
-            AdapterUtils.setErrorResponse(response, ErrorCode.GET_VARIABLE_FAILURE,
-                    String.format("Failed to get variables. Reason: '%s'", e.getMessage()));
-        }
-    }
-
-    private void setVariable(Requests.SetVariableArguments arguments, Response response) {
-        response.body = this.variableRequestHandler.setVariable(arguments);
-    }
-
-    private void evaluate(Requests.EvaluateArguments arguments, Response response) {
-        try {
-            response.body = this.variableRequestHandler.evaluate(arguments);
-        } catch (IllegalArgumentException | NullPointerException ex) {
-            AdapterUtils.setErrorResponse(response, ErrorCode.EVALUATE_FAILURE,
-                    String.format("Failed to evaluate expression %s . Reason: '%s'", arguments.expression, ex.toString()));
-        }
-    }
-
-    /* ======================================================*/
-    /* Dispatch logic End */
-    /* ======================================================*/
-
-    private class VariableRequestHandler {
-        private static final String PATTERN = "([a-zA-Z_0-9$]+)\\s*\\(([^)]+)\\)";
-        private final Pattern simpleExprPattern = Pattern.compile("[A-Za-z0-9_.\\s]+");
-        private IVariableFormatter variableFormatter;
-
-        public VariableRequestHandler(IVariableFormatter variableFormatter) {
-            this.variableFormatter = variableFormatter;
-        }
-
-        Responses.ResponseBody scopes(Requests.ScopesArguments arguments) {
-            List<Types.Scope> scopes = new ArrayList<>();
-            JdiObjectProxy<StackFrame> stackFrameProxy = (JdiObjectProxy<StackFrame>) debugContext.getRecyclableIdPool().getObjectById(arguments.frameId);
-            if (stackFrameProxy == null) {
-                return new Responses.ScopesResponseBody(scopes);
-            }
-            StackFrameScope localScope = new StackFrameScope(stackFrameProxy.getProxiedObject(), "Local");
-            scopes.add(new Types.Scope(
-                    localScope.getScope(), debugContext.getRecyclableIdPool().addObject(stackFrameProxy.getProxiedObject()
-                    .thread().uniqueID(), localScope), false));
-
-            return new Responses.ScopesResponseBody(scopes);
-        }
-
-
-        Responses.ResponseBody variables(Requests.VariablesArguments arguments) throws AbsentInformationException {
-            Map<String, Object> options = variableFormatter.getDefaultOptions();
-            // This should be false by default(currently true for test).
-            // User will need to explicitly turn it on by configuring launch.json
-            boolean showStaticVariables = true;
-            // TODO: when vscode protocol support customize settings of value format, showFullyQualifiedNames should be one of the options.
-            boolean showFullyQualifiedNames = true;
-            if (arguments.format != null && arguments.format.hex) {
-                options.put(NumericFormatter.NUMERIC_FORMAT_OPTION, NumericFormatEnum.HEX);
-            }
-            if (showFullyQualifiedNames) {
-                options.put(SimpleTypeFormatter.QUALIFIED_CLASS_NAME_OPTION, showFullyQualifiedNames);
-            }
-
-            List<Types.Variable> list = new ArrayList<>();
-            List<Variable> variables;
-            Object obj = debugContext.getRecyclableIdPool().getObjectById(arguments.variablesReference);
-            // vscode will always send variables request to a staled scope, return the empty list is ok since the next
-            // variable request will contain the right variablesReference.
-            if (obj == null) {
-                return new Responses.VariablesResponseBody(list);
-            }
-            ThreadReference thread;
-            if (obj instanceof StackFrameScope) {
-                StackFrame frame = ((StackFrameScope) obj).getStackFrame();
-                thread = frame.thread();
-                variables = VariableUtils.listLocalVariables(frame);
-                Variable thisVariable = VariableUtils.getThisVariable(frame);
-                if (thisVariable != null) {
-                    variables.add(thisVariable);
-                }
-                if (showStaticVariables && frame.location().method().isStatic()) {
-                    variables.addAll(VariableUtils.listStaticVariables(frame));
-                }
-            } else if (obj instanceof ThreadObjectReference) {
-                ObjectReference currentObj = ((ThreadObjectReference) obj).getObject();
-                thread = ((ThreadObjectReference) obj).getThread();
-
-                if (arguments.count > 0) {
-                    variables = VariableUtils.listFieldVariables(currentObj, arguments.start, arguments.count);
-                } else {
-                    variables = VariableUtils.listFieldVariables(currentObj, showStaticVariables);
-                }
-
-            } else {
-                throw new IllegalArgumentException(String
-                        .format("VariablesRequest: Invalid variablesReference %d.", arguments.variablesReference));
-            }
-            // find variable name duplicates
-            Set<String> duplicateNames = getDuplicateNames(variables.stream().map(var -> var.name)
-                    .collect(Collectors.toList()));
-            Map<Variable, String> variableNameMap = new HashMap<>();
-            if (!duplicateNames.isEmpty()) {
-                Map<String, List<Variable>> duplicateVars =
-                        variables.stream()
-                                .filter(var -> duplicateNames.contains(var.name))
-                                .collect(Collectors.groupingBy(var -> var.name, Collectors.toList()));
-
-                duplicateVars.forEach((k, duplicateVariables) -> {
-                    Set<String> declarationTypeNames = new HashSet<>();
-                    boolean declarationTypeNameConflict = false;
-                    // try use type formatter to resolve name conflict
-                    for (Variable javaVariable : duplicateVariables) {
-                        Type declarationType = javaVariable.getDeclaringType();
-                        if (declarationType != null) {
-                            String declarationTypeName = this.variableFormatter.typeToString(declarationType, options);
-                            String compositeName = String.format("%s (%s)", javaVariable.name, declarationTypeName);
-                            if (!declarationTypeNames.add(compositeName)) {
-                                declarationTypeNameConflict = true;
-                                break;
-                            }
-                            variableNameMap.put(javaVariable, compositeName);
-                        }
-                    }
-                    // if there are duplicate names on declaration types, use fully qualified name
-                    if (declarationTypeNameConflict) {
-                        for (Variable javaVariable : duplicateVariables) {
-                            Type declarationType = javaVariable.getDeclaringType();
-                            if (declarationType != null) {
-                                variableNameMap.put(javaVariable, String.format("%s (%s)", javaVariable.name, declarationType.name()));
-                            }
-                        }
-                    }
-                });
-            }
-            for (Variable javaVariable : variables) {
-                Value value = javaVariable.value;
-                String name = javaVariable.name;
-                if (variableNameMap.containsKey(javaVariable)) {
-                    name = variableNameMap.get(javaVariable);
-                }
-                int referenceId = 0;
-                if (value instanceof ObjectReference && VariableUtils.hasChildren(value, showStaticVariables)) {
-                    ThreadObjectReference threadObjectReference = new ThreadObjectReference(thread, (ObjectReference) value);
-                    referenceId = debugContext.getRecyclableIdPool().addObject(thread.uniqueID(), threadObjectReference);
-                }
-                Types.Variable typedVariables = new Types.Variable(name, variableFormatter.valueToString(value, options),
-                        variableFormatter.typeToString(value == null ? null : value.type(), options), referenceId, null);
-                if (javaVariable.value instanceof ArrayReference) {
-                    typedVariables.indexedVariables = ((ArrayReference) javaVariable.value).length();
-                }
-                list.add(typedVariables);
-            }
-            return new Responses.VariablesResponseBody(list);
-        }
-
-        Responses.ResponseBody setVariable(Requests.SetVariableArguments arguments) {
-            Map<String, Object> options = variableFormatter.getDefaultOptions();
-            // This should be false by default(currently true for test).
-            // User will need to explicitly turn it on by configuring launch.json
-            boolean showStaticVariables = true;
-            // TODO: when vscode protocol support customize settings of value format, showFullyQualifiedNames should be one of the options.
-            boolean showFullyQualifiedNames = true;
-            if (arguments.format != null && arguments.format.hex) {
-                options.put(NumericFormatter.NUMERIC_FORMAT_OPTION, NumericFormatEnum.HEX);
-            }
-            if (showFullyQualifiedNames) {
-                options.put(SimpleTypeFormatter.QUALIFIED_CLASS_NAME_OPTION, showFullyQualifiedNames);
-            }
-
-            Object obj = debugContext.getRecyclableIdPool().getObjectById(arguments.variablesReference);
-
-            // obj is null means the stackframe is continued by user manually,
-            if (obj == null) {
-                return new Responses.ErrorResponseBody(
-                        new Types.Message(ErrorCode.SET_VARIABLE_FAILURE.getId(), "Cannot set value because the thread is resumed."));
-            }
-            ThreadReference thread;
-            String name = arguments.name;
-            Value newValue;
-            String belongToClass = null;
-
-            if (arguments.name.contains("(")) {
-                name = arguments.name.replaceFirst(PATTERN, "$1");
-                belongToClass = arguments.name.replaceFirst(PATTERN, "$2");
-            }
-
-            try {
-                if (obj instanceof StackFrameScope) {
-                    StackFrameScope frameScope = (StackFrameScope) obj;
-                    thread = frameScope.getStackFrame().thread();
-                    newValue = handleSetValueForStackFrame(name, belongToClass, arguments.value,
-                            showStaticVariables, frameScope.getStackFrame(), options);
-                } else if (obj instanceof ThreadObjectReference) {
-                    ObjectReference currentObj = ((ThreadObjectReference) obj).getObject();
-                    thread = ((ThreadObjectReference) obj).getThread();
-                    newValue = handleSetValueForObject(name, belongToClass, arguments.value,
-                            currentObj, options);
-                } else {
-                    throw new IllegalArgumentException(
-                            String.format("SetVariableRequest: Variable %s cannot be found.", arguments.variablesReference));
-                }
-            } catch (IllegalArgumentException | AbsentInformationException | InvalidTypeException
-                    | UnsupportedOperationException | ClassNotLoadedException e) {
-                return new Responses.ErrorResponseBody(new Types.Message(ErrorCode.SET_VARIABLE_FAILURE.getId(), e.toString()));
-            }
-            int referenceId = getReferenceId(thread, newValue, showStaticVariables);
-
-            int indexedVariables = 0;
-            if (newValue instanceof ArrayReference) {
-                indexedVariables = ((ArrayReference) newValue).length();
-            }
-            return new Responses.SetVariablesResponseBody(
-                    this.variableFormatter.typeToString(newValue == null ? null : newValue.type(), options), // type
-                    this.variableFormatter.valueToString(newValue, options), // value,
-                    referenceId, indexedVariables);
-
-        }
-
-        private Value handleSetValueForObject(String name, String belongToClass, String valueString,
-                                              ObjectReference currentObj, Map<String, Object> options)
-                throws InvalidTypeException, ClassNotLoadedException {
-            Value newValue;
-            if (currentObj instanceof ArrayReference) {
-                ArrayReference array = (ArrayReference) currentObj;
-                Type eleType = ((ArrayType) array.referenceType()).componentType();
-                newValue = setArrayValue(array, eleType, Integer.parseInt(name), valueString, options);
-            } else {
-                if (StringUtils.isBlank(belongToClass)) {
-                    Field field = currentObj.referenceType().fieldByName(name);
-                    if (field != null) {
-                        if (field.isStatic()) {
-                            newValue = this.setStaticFieldValue(currentObj.referenceType(), field,
-                                    name, valueString, options);
-                        } else {
-                            newValue = this.setObjectFieldValue(currentObj, field, name,
-                                    valueString, options);
-                        }
-                    } else {
-                        throw new IllegalArgumentException(
-                                String.format("SetVariableRequest: Variable %s cannot be found.", name));
-                    }
-                } else {
-                    newValue = setFieldValueWithConflict(currentObj, currentObj.referenceType().allFields(),
-                            name, belongToClass, valueString, options);
-                }
-            }
-            return newValue;
-
-        }
-
-        private Value handleSetValueForStackFrame(String name, String belongToClass, String valueString,
-                                                  boolean showStaticVariables,
-                                                  StackFrame frame, Map<String, Object> options)
-                throws AbsentInformationException, InvalidTypeException, ClassNotLoadedException {
-            Value newValue;
-            if (name.equals("this")) {
-                throw new UnsupportedOperationException("SetVariableRequest: 'This' variable cannot be changed.");
-            }
-            LocalVariable variable = frame.visibleVariableByName(name);
-            if (StringUtils.isBlank(belongToClass) && variable != null) {
-                newValue = this.setFrameValue(frame, variable, valueString, options);
-            } else {
-                if (showStaticVariables && frame.location().method().isStatic()) {
-                    ReferenceType type = frame.location().declaringType();
-                    if (StringUtils.isBlank(belongToClass)) {
-                        Field field = type.fieldByName(name);
-                        newValue = setStaticFieldValue(type, field, name, valueString, options);
-                    } else {
-                        newValue = setFieldValueWithConflict(null, type.allFields(), name, belongToClass,
-                                valueString, options);
-                    }
-
-                } else {
-                    throw new UnsupportedOperationException(
-                            String.format("SetVariableRequest: Variable %s cannot be found.", name));
-                }
-            }
-            return newValue;
-        }
-
-        private Responses.ResponseBody evaluate(Requests.EvaluateArguments arguments) {
-            // This should be false by default(currently true for test).
-            // User will need to explicitly turn it on by configuring launch.json
-            final boolean showStaticVariables = true;
-            // TODO: when vscode protocol support customize settings of value format, showFullyQualifiedNames should be one of the options.
-            boolean showFullyQualifiedNames = true;
-            Map<String, Object> options = variableFormatter.getDefaultOptions();
-            if (arguments.format != null && arguments.format.hex) {
-                options.put(NumericFormatter.NUMERIC_FORMAT_OPTION, NumericFormatEnum.HEX);
-            }
-            if (showFullyQualifiedNames) {
-                options.put(SimpleTypeFormatter.QUALIFIED_CLASS_NAME_OPTION, showFullyQualifiedNames);
-            }
-            String expression = arguments.expression;
-
-            if (StringUtils.isBlank(expression)) {
-                throw new IllegalArgumentException("Empty expression cannot be evaluated.");
-            }
-
-            if (!simpleExprPattern.matcher(expression).matches()) {
-                throw new IllegalArgumentException("Complicate expression is not supported currently.");
-            }
-
-            JdiObjectProxy<StackFrame> stackFrameProxy = (JdiObjectProxy<StackFrame>) debugContext.getRecyclableIdPool().getObjectById(arguments.frameId);
-            if (stackFrameProxy == null) {
-                // stackFrameProxy is null means the stackframe is continued by user manually,
-                return new Responses.ErrorResponseBody(
-                        new Types.Message(ErrorCode.EVALUATE_FAILURE.getId(), "Cannot evaluate because the thread is resumed."));
-            }
-
-            // split a.b.c => ["a", "b", "c"]
-            List<String> referenceExpressions = Arrays.stream(StringUtils.split(expression, '.'))
-                    .filter(StringUtils::isNotBlank).map(StringUtils::trim).collect(Collectors.toList());
-
-            // get first level of value from stack frame
-            Variable firstLevelValue = null;
-            boolean inStaticMethod = !stackFrameProxy.getProxiedObject().location().method().isStatic();
-            String firstExpression = referenceExpressions.get(0);
-            // handle special case of 'this'
-            if (firstExpression.equals("this") && !inStaticMethod) {
-                firstLevelValue = VariableUtils.getThisVariable(stackFrameProxy.getProxiedObject());
-            }
-            if (firstLevelValue == null) {
-                try {
-                    // local variables first, that means
-                    // if both local variable and static variable are found, use local variable
-                    List<Variable> localVariables = VariableUtils.listLocalVariables(stackFrameProxy.getProxiedObject());
-                    List<Variable> matchedLocal = localVariables.stream()
-                            .filter(localVariable -> localVariable.name.equals(firstExpression)).collect(Collectors.toList());
-                    if (!matchedLocal.isEmpty()) {
-                        firstLevelValue = matchedLocal.get(0);
-                    } else {
-                        List<Variable> staticVariables = VariableUtils.listStaticVariables(stackFrameProxy.getProxiedObject());
-                        List<Variable> matchedStatic = staticVariables.stream()
-                                .filter(staticVariable -> staticVariable.name.equals(firstExpression)).collect(Collectors.toList());
-                        if (matchedStatic.isEmpty()) {
-                            throw new IllegalArgumentException(String.format("Cannot find the variable: %s.", referenceExpressions.get(0)));
-                        }
-                        firstLevelValue = matchedStatic.get(0);
-                    }
-
-                } catch (AbsentInformationException e) {
-                    // ignore
-                }
-            }
-
-            if (firstLevelValue == null) {
-                throw new IllegalArgumentException(String.format("Cannot find variable with name '%s'.", referenceExpressions.get(0)));
-            }
-            ThreadReference thread = stackFrameProxy.getProxiedObject().thread();
-            Value currentValue = firstLevelValue.value;
-
-            for (int i = 1; i < referenceExpressions.size(); i++) {
-                String fieldName = referenceExpressions.get(i);
-                if (currentValue == null) {
-                    throw new NullPointerException("Evaluation encounters NPE error.");
-                }
-                if (currentValue instanceof PrimitiveValue) {
-                    throw new IllegalArgumentException(String.format("Cannot find the field: %s.", fieldName));
-                }
-                if (currentValue instanceof ArrayReference) {
-                    throw new IllegalArgumentException(String.format("Evaluating array elements is not supported currently.", fieldName));
-                }
-                ObjectReference obj = (ObjectReference) currentValue;
-                Field field = obj.referenceType().fieldByName(fieldName);
-                if (field == null) {
-                    throw new IllegalArgumentException(String.format("Cannot find the field: %s.", fieldName));
-                }
-                if (field.isStatic()) {
-                    throw new IllegalArgumentException(String.format("Cannot find the field: %s.", fieldName));
-                }
-                currentValue = obj.getValue(field);
-            }
-
-            int referenceId = 0;
-            if (currentValue instanceof ObjectReference && VariableUtils.hasChildren(currentValue, showStaticVariables)) {
-                // save the evaluated value in object pool, because like java.lang.String, the evaluated object will have sub structures
-                // we need to set up the id map.
-                ThreadObjectReference threadObjectReference = new ThreadObjectReference(thread, (ObjectReference) currentValue);
-                referenceId = debugContext.getRecyclableIdPool().addObject(thread.uniqueID(), threadObjectReference);
-            }
-            int indexedVariables = 0;
-            if (currentValue instanceof ArrayReference) {
-                indexedVariables = ((ArrayReference) currentValue).length();
-            }
-            return new Responses.EvaluateResponseBody(variableFormatter.valueToString(currentValue, options),
-                    referenceId, variableFormatter.typeToString(currentValue == null ? null : currentValue.type(), options),
-                    indexedVariables);
-        }
-
-        private Value setValueProxy(Type type, String value, SetValueFunction setValueFunc, Map<String, Object> options)
-                throws ClassNotLoadedException, InvalidTypeException {
-            Value newValue = this.variableFormatter.stringToValue(value, type, options);
-            setValueFunc.apply(newValue);
-            return newValue;
-        }
-
-        private Value setStaticFieldValue(Type declaringType, Field field, String name, String value, Map<String, Object> options)
-                throws ClassNotLoadedException, InvalidTypeException {
-            if (field.isFinal()) {
-                throw new UnsupportedOperationException(
-                        String.format("SetVariableRequest: Final field %s cannot be changed.", name));
-            }
-            if (!(declaringType instanceof ClassType)) {
-                throw new UnsupportedOperationException(
-                        String.format("SetVariableRequest: Field %s in interface cannot be changed.", name));
-            }
-            return setValueProxy(field.type(), value, newValue -> ((ClassType) declaringType).setValue(field, newValue), options);
-        }
-
-        private Value setFrameValue(StackFrame frame, LocalVariable localVariable, String value, Map<String, Object> options)
-                throws ClassNotLoadedException, InvalidTypeException {
-            return setValueProxy(localVariable.type(), value, newValue -> frame.setValue(localVariable, newValue), options);
-        }
-
-        private Value setObjectFieldValue(ObjectReference obj, Field field, String name, String value, Map<String, Object> options)
-                throws ClassNotLoadedException, InvalidTypeException {
-            if (field.isFinal()) {
-                throw new UnsupportedOperationException(
-                        String.format("SetVariableRequest: Final field %s cannot be changed.", name));
-            }
-            return setValueProxy(field.type(), value, newValue -> obj.setValue(field, newValue), options);
-        }
-
-        private Value setArrayValue(ArrayReference array, Type eleType, int index, String value, Map<String, Object> options)
-                throws ClassNotLoadedException, InvalidTypeException {
-            return setValueProxy(eleType, value, newValue -> array.setValue(index, newValue), options);
-        }
-
-        private Value setFieldValueWithConflict(ObjectReference obj, List<Field> fields, String name, String belongToClass,
-                                                String value, Map<String, Object> options) throws ClassNotLoadedException, InvalidTypeException {
-            Field field;
-            // first try to resolve field by fully qualified name
-            List<Field> narrowedFields = fields.stream().filter(TypeComponent::isStatic)
-                    .filter(t -> t.name().equals(name) && t.declaringType().name().equals(belongToClass))
-                    .collect(Collectors.toList());
-            if (narrowedFields.isEmpty()) {
-                // second try to resolve field by formatted name
-                narrowedFields = fields.stream().filter(TypeComponent::isStatic)
-                        .filter(t -> t.name().equals(name)
-                                && this.variableFormatter.typeToString(t.declaringType(), options).equals(belongToClass))
-                        .collect(Collectors.toList());
-            }
-            if (narrowedFields.size() == 1) {
-                field = narrowedFields.get(0);
-            } else {
-                throw new UnsupportedOperationException(String.format("SetVariableRequest: Name conflicted for %s.", name));
-            }
-            return field.isStatic() ? setStaticFieldValue(field.declaringType(), field, name, value, options)
-                    : this.setObjectFieldValue(obj, field, name, value, options);
-
-        }
-
-        private int getReferenceId(ThreadReference thread, Value value, boolean includeStatic) {
-            if (value instanceof ObjectReference && VariableUtils.hasChildren(value, includeStatic)) {
-                ThreadObjectReference threadObjectReference = new ThreadObjectReference(thread, (ObjectReference) value);
-                return debugContext.getRecyclableIdPool().addObject(thread.uniqueID(), threadObjectReference);
-            }
-            return 0;
-        }
-
-
-        private Set<String> getDuplicateNames(Collection<String> list) {
-            Set<String> result = new HashSet<>();
-            Set<String> set = new HashSet<>();
-
-            for (String item : list) {
-                if (!set.contains(item)) {
-                    set.add(item);
-                } else {
-                    result.add(item);
-                }
-            }
-            return result;
-        }
-    }
-
-    @FunctionalInterface
-    interface SetValueFunction {
-        void apply(Value value) throws InvalidTypeException, ClassNotLoadedException;
     }
 }
