+++ org.eclipse.jdt.ls.debug/src/org/eclipse/jdt/ls/debug/adapter/handler/InitializeRequestHandler.java
@@ -49,6 +49,7 @@ public class InitializeRequestHandler implements IDebugRequestHandler {
         Types.Capabilities caps = new Types.Capabilities();
         caps.supportsConfigurationDoneRequest = true;
         caps.supportsHitConditionalBreakpoints = true;
+        caps.supportsSetVariable = true;
         caps.supportTerminateDebuggee = true;
         Types.ExceptionBreakpointFilter[] exceptionFilters = {
                 Types.ExceptionBreakpointFilter.UNCAUGHT_EXCEPTION_FILTER,
